마커 삭제 오류 수정
